add folder routes,endPoints and modules and change Data folder name to modules and add folder to database
