+++ Modules/server_instance.py
@@ -1,0 +1,3 @@
+from Modules.server_function import serverFunction
+
+server_function = serverFunction()
